Rename singleThreadGAforSGA to UnimproveStop
修改為正確名稱
