+++ src/openga/applications/singleMachineOAS_SGA.java
@@ -71,7 +71,7 @@ public class singleMachineOAS_SGA extends mTSPSGATwoPart {
 
   @Override
   public void initiateVars() {
-    GaMain = new singleThreadGAforSGA();
+    GaMain = new singleThreadGAUnimproveStop();
     Population = new population();
     Selection = new binaryTournament();
     Crossover = new PMX();
@@ -104,19 +104,19 @@ public class singleMachineOAS_SGA extends mTSPSGATwoPart {
     String implementResult = instanceName + "\t" + DEFAULT_crossoverRate + "\t" + DEFAULT_mutationRate + "\t" + DEFAULT_PopSize + "\t" + applyLocalSearch + "\t" + alpha
             + "\t" + GaMain.getArchieve().getSingleChromosome(0).getObjValue()[0]
             + "\t" + timeClock1.getExecutionTime() / 1000.0 + "\n";
-    writeFile("OASforSMSP_20190215-SGA" + "MaxRevenueFull", implementResult);
+    writeFile("OASforSMSP_20190219-SGA" + "MaxRevenueFull", implementResult);
     System.out.print(implementResult);
   }
 
   public static void main(String[] args) {
-    System.out.println("OASforSMSP_20190215" + "MaxRevenueFull");
+    System.out.println("OASforSMSP_20190219" + "MaxRevenueFull");
     int counter = 0;
     boolean applyLocalSearch;
     double[] crossoverRate = new double[]{0.5};//0.9, 0.5 [0.5]
     double[] mutationRate = new double[]{0.5};//0.1, 0.5 [0.5]
     double elitism[] = new double[]{0.1};
     int[] crossoverType = new int[]{2};//0: All salesmen reserve the same sites,2: Last salesmen reserve the same sites,3: TCX (Original) [2]
-    int repeat = 1;//30
+    int repeat = 30;//30
     int generations[] = new int[]{0};//1000
     int[] populationsSize = new int[]{60};//50, 100, 200 [50]
     int[] numberOfSalesmen = new int[]{2};
